Fixing unittests when running tests weren't running on a clean system
Some of our unittests used a temporary directory to store job definitions while testing. These tests didn't clean up the directory on tear down. When the second test ran it asserted that there weren't any jobs stored there, but it found the jobs from the previous test. As the directory is now deleted on tearDown the assert won't fail.

Another way to circumvent this would have been to create directories with unique names in setUp. This would have also left artifacts in the temporary directory if we wouldn't have cleaned that up either. I don't expect our tests to be run in parralel when the unique approach would prevent collisions.