Add fetcher field to Chronos jobs and improve future compatibility
As the uris field is being deprecated we added the fetcher field which is
replacing the uris functionality.

We also store fields in the Chronos job definitions that we don't know and not
just drop them as we previously did. This will allow users to use fields in
their job definitions that we don't know about without silently breaking their
jobs when commiting to Chronos. Also this is much closer to the expected
behaviour of chapi being a nicer `curl` replacement with _additional_ features.