Reverting "future compatibility"
It would have been nice to not throw away fields we don't know about yet
and just pass them on to the server that can than tell the users that
their job definitions are broken.

But the assumption that every field in the iterables returned by
`getSimpleArrayCopy()`, `jsonSerialize()` and `getIterator()` is a
property of the entity is so deeply infused in chapi. This also crops up
in weird places I didn't expect it would.

Because these changes would've ended up to be quite large and we're
likely to phase out the use of chapi I decided against implementing
them. So this is reverting the changes I've already made.