Remove uris field from Chronos jobs
Chronos converts from uris to fetch and from fetch to uris so that both
fields are populated with the same data. This leads to diffs when you
only use one field in your *.json files. As we had problems with uris
and fetch is more expressive we dropped support for uris.