Unit tests: Remove tests testGetJobsSuccess
This test is removed, because it doesn't do any assertion.
We (trivago) nearly don't use Marathon anymore, hence there is no big need
for the Marathon integration from chapi. From the community, we are not
aware that anyone is using Marathon in combination with chapi.

Hence, we are removing this test and don't put a lot of effort
in maintaining the marathon-chapi state